update web and mobile

+++ urbano/controllers/PublicacaopesqController.php
@@ -2,12 +2,14 @@
 
 namespace app\controllers;
 
+use app\models\Avaliacaopubpesq;
 use Yii;
 use app\models\Publicacaopesq;
 use app\models\PublicacaopesqSearch;
 use yii\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\web\UploadedFile;
 
 /**
  * PublicacaopesqController implements the CRUD actions for Publicacaopesq model.
@@ -60,17 +62,62 @@ class PublicacaopesqController extends Controller
         ]);
     }
 
+    public function getphotos($model){
+        $path1 =  UploadedFile::getInstance($model, 'img1');
+        $path2 =  UploadedFile::getInstance($model, 'img2');
+        $path3 =  UploadedFile::getInstance($model, 'img3');
+        $path4 =  UploadedFile::getInstance($model, 'img4');
+
+        //$type = pathinfo($path, PATHINFO_EXTENSION);
+        if($path1 != null && $path1->tempName != null) {
+            $data1 = file_get_contents($path1->tempName);
+            $img1 = base64_encode($data1);
+            $model->img1 = (string)$img1;
+        }
+        if($path2 != null && $path2->tempName != null) {
+            $data2 = file_get_contents($path2->tempName);
+            $img2 = base64_encode($data2);
+            $model->img2 = (string)$img2;
+        }
+        if($path3 != null && $path3->tempName != null) {
+            $data3 = file_get_contents($path3->tempName);
+            $img3 = base64_encode($data3);
+            $model->img3 = (string)$img3;
+        }
+        if($path4 != null && $path4->tempName != null) {
+            $data4 = file_get_contents($path4->tempName);
+            $img4 = base64_encode($data4);
+            $model->img4 = (string)$img4;
+        }
+        return $model;
+    }
+
+
+
     /**
      * Creates a new Publicacaopesq model.
      * If creation is successful, the browser will be redirected to the 'view' page.
      * @return mixed
      */
-    public function actionCreate()
-    {
+    public function actionCreate(){
         $model = new Publicacaopesq();
 
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+        $model->latitude = -16.3364897;
+        $model->longitude = -48.9413023;
+
+
+        if ($model->load(Yii::$app->request->post())) {
+            if(Yii::$app->request->isPost){
+                $model =$this->getphotos($model);
+                //$model->img2 = 'data:image/' . $type . ';base64,' . base64_encode($data);
+            }
+            if($model->save()){
+                return $this->redirect(['view', 'id' => $model->id]);
+            }else{
+                return $this->render('create', [
+                    'model' => $model,
+                ]);
+            }
         }
 
         return $this->render('create', [
@@ -88,9 +135,17 @@ class PublicacaopesqController extends Controller
     public function actionUpdate($id)
     {
         $model = $this->findModel($id);
-
-        if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['view', 'id' => $model->id]);
+        if ($model->load(Yii::$app->request->post())) {
+            if(Yii::$app->request->isPost){
+                $model =$this->getphotos($model);
+            }
+            if($model->save()){
+                return $this->redirect(['view', 'id' => $model->id]);
+            }else{
+                return $this->render('update', [
+                    'model' => $model,
+                ]);
+            }
         }
 
         return $this->render('update', [
@@ -107,6 +162,11 @@ class PublicacaopesqController extends Controller
      */
     public function actionDelete($id)
     {
+        if (Avaliacaopubpesq::find()->where(['idpubpesq'=>$id])->all()){
+            Yii::$app->session->setFlash('danger', "Esta publicação contem avaliação");
+            return $this->redirect(['index']);
+        }
+
         $this->findModel($id)->delete();
 
         return $this->redirect(['index']);
